backend/requirements.txt
fastapi
uvicorn
websockets
sqlmodel
psycopg2-binary
pandas
dotenv
openai
ta
asyncpg
bcrypt
cryptography
python-jose
python-multipart
python-dotenv
pydantic
pydantic-settings
pydantic[email]
python-decouple
jwt
requests